requirements.txt
# Core LLM / LangChain
langchain>=0.3.0
langchain-ollama>=0.2.0
langchain-community>=0.3.0
langchain-chroma>=0.1.4
langgraph>=0.2.0
langchain-core>=0.3.0

# Vector Store
chromadb>=0.5.0

# Frontend
streamlit>=1.39.0

# Data Processing
pyyaml>=6.0
requests>=2.32.0
gitpython>=3.1.0
pandas>=2.2.0
numpy>=1.26.0

# HuggingFace datasets (for datasets 4, 5, 6)
datasets>=2.14.0

# Visualization
plotly>=5.18.0
umap-learn>=0.5.6
scikit-learn>=1.3.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
